+++ CHANGELOG.md
@@ -2,3 +2,6 @@
 # Version
 - Remove `From<String>` impl for `ToKserdErr`
 - Cleaned up error handling
+- Implement `ToKserd` for tuples up to 12 length
+- Implement `ToKserd` for `Box<T>`
+- Implement `ToKserd` for `Option<T>`